Fix an issue where keyChannel for loop hangs
Because we cannot know how many suggestions there will be prior to
starting the search, we need to keep some sort of a count register
and terminate the keyChannel suggestions for loop when no more
suggestions are found.

This change passes a context structre to each requestKeyWords() goroutine
and uses mutexes to sync the inc/dec of the total keywords found count.